keep things DRY, move identical methods to base classes
